refactor: make fetchLogsForBatch reusuable
